# openapi.yaml
openapi: 3.0.3
info:
  title: Aframp Pay API
  version: 0.1.0
  description: >
    Merchant-facing backend for Aframp Pay — Stellar-settled point-of-sale payments.
    Prose reference with frontend guidance lives in API.md; this spec is the
    machine-readable contract for generating typed clients.


    Money is always integer stroops (1 unit = 10,000,000 stroops), never a float.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /compliance/reports:
    post:
      tags: [Compliance]
      summary: Generate compliance report
      description: Generate a new compliance effectiveness report.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateReportRequest' }
      responses:
        '201':
          description: Report generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComplianceReport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Compliance]
      summary: List compliance reports
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Reports list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportListPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /compliance/reports/{id}:
    get:
      tags: [Compliance]
      summary: Get compliance report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Compliance report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComplianceReport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256, 24h expiry. No refresh endpoint — a 401 means re-login. For API
        clients and scripts; browsers should use cookieAuth instead.
    cookieAuth:
      type: apiKey
      in: cookie
      name: aframp_session
      description: >
        HttpOnly session cookie set by /signup and /login, cleared by /logout.
        The preferred method for browsers — JS cannot read it, so an XSS on the
        page cannot steal the session. Never copy the token into localStorage.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Max rows to return. Clamped server-side to 1-200.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: Validation failed, or the account has no merchant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing, malformed, or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, example: insufficient available balance }
...
    GenerateReportRequest:
      type: object
      required: [report_type, period_start, period_end, format]
      properties:
        report_type:
          type: string
          enum: [monthly, quarterly, annual, ad_hoc]
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        format:
          type: string
          enum: [pdf, csv, json]
    ComplianceReport:
      type: object
      required: [id, report_type, period_start, period_end, generated_by, generated_at, format]
      properties:
        id: { type: string, format: uuid }
        report_type:
          type: string
          enum: [monthly, quarterly, annual, ad_hoc]
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        total_alerts: { type: integer }
        sanctions_alerts: { type: integer }
        aml_alerts: { type: integer }
        kyc_alerts: { type: integer }
        false_positives: { type: integer }
        false_positive_rate: { type: number }
        avg_resolution_time_hrs: { type: number }
        median_resolution_time_hrs: { type: number }
        sla_breaches: { type: integer }
        sla_compliance_rate: { type: number }
        cases_cleared: { type: integer }
        cases_blocked: { type: integer }
        cases_pending: { type: integer }
        low_risk_cases: { type: integer }
        medium_risk_cases: { type: integer }
        critical_risk_cases: { type: integer }
        alert_volume_trend: { type: string, nullable: true }
        false_positive_trend: { type: string, nullable: true }
        generated_by: { type: string }
        generated_at: { type: string, format: date-time }
        format:
          type: string
          enum: [pdf, csv, json]
        file_path: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    ReportListPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ComplianceReport' }
        total: { type: integer }